# backend/app/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ChronoSift API",
    "description": "Forensic Timeline Analysis API - Comprehensive documentation for all endpoints",
    "version": "1.0.0",
    "contact": {
      "name": "ChronoSift",
      "url": "https://github.com/yourusername/chronosift"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Development server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from /api/auth/login"
      }
    },
    "schemas": {
...
      "Project": {
        "type": "object",
        "properties": {
          "id": {"type": "integer"},
          "name": {"type": "string"},
          "description": {"type": "string"},
          "team_id": {"type": "integer"},
          "status": {"type": "string"},
          "created_at": {"type": "string", "format": "date-time"}
        }
      },
...
    }
  },
  "paths": {
...
    "/projects": {
      "get": {
        "tags": ["Projects"],
        "summary": "List all projects user has access to",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {
            "name": "team_id",
            "in": "query",
            "schema": {"type": "integer"},
            "description": "Filter by team ID"
          }
        ],
        "responses": {
          "200": {
            "description": "List of projects",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {"$ref": "#/components/schemas/Project"}
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Projects"],
        "summary": "Create a new project",
        "security": [{"BearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "team_id"],
                "properties": {
                  "name": {"type": "string"},
                  "description": {"type": "string"},
                  "team_id": {"type": "integer"},
                  "status": {"type": "string", "default": "active"}
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Project created",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/Project"}
              }
            }
          }
        }
      }
    },
    "/projects/{id}/search": {
      "get": {
        "tags": ["Projects"],
        "summary": "Search across all timelines in a project",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {"type": "integer"}
          },
          {
            "name": "keyword",
            "in": "query",
            "schema": {"type": "string"}
          },
          {
            "name": "start_date",
            "in": "query",
            "schema": {"type": "string", "format": "date-time"}
          },
          {
            "name": "end_date",
            "in": "query",
            "schema": {"type": "string", "format": "date-time"}
          },
          {
            "name": "timeline_ids",
            "in": "query",
            "schema": {"type": "array", "items": {"type": "integer"}}
          }
        ],
        "responses": {
          "200": {
            "description": "Search results"
          }
        }
      }
    },
...
  }
}